==> notebooks/0.embed.ipynb <==
import torch
import polars as pl
from tqdm.auto import tqdm  # Add this at top if not already
import pickle


PATH = "D:/UMLS/"
FILE_PATH = "D:/UMLS/exact_mapping.csv"
SAVE_PATH = "D:/UMLS/icd_snomed_mappings.pkl"



# Load the CSV file into a Polars DataFrame
df_mapping = pl.read_csv(FILE_PATH)

df_mapping_snomed = df_mapping["SNOMED_ID", "SNOMED_label"].unique()
s = pl.Series("idx", range(len(df_mapping_snomed)))
df_mapping_snomed.insert_column(0, s).write_csv("D:/UMLS/snomed_info.csv")

df_mapping_icd = df_mapping["ICD_ID", "ICD_label"].unique()
s = pl.Series("idx", range(len(df_mapping_icd)))
df_mapping_icd.insert_column(0, s).write_csv("D:/UMLS/icd_info.csv")

id2snomed = dict(zip(df_mapping_snomed["idx"], df_mapping_snomed["SNOMED_ID"]))
id2icd = dict(zip(df_mapping_icd["idx"], df_mapping_icd["ICD_ID"]))
snomed2id = dict(zip(df_mapping_snomed["SNOMED_ID"], df_mapping_snomed["idx"]))
icd2id = dict(zip(df_mapping_icd["ICD_ID"], df_mapping_icd["idx"]))

df_mapping = df_mapping.with_columns(idx_snomed = pl.col("SNOMED_ID").replace(snomed2id),
                        idx_icd = pl.col("ICD_ID").replace(icd2id))

df_icd2snomed = df_mapping.group_by("idx_icd").agg("idx_snomed")
icd2snomed = dict(zip(df_icd2snomed["idx_icd"], df_icd2snomed["idx_snomed"]))

df_snomed2icd = df_mapping.group_by("idx_snomed").agg("idx_icd")
snomed2icd = dict(zip(df_snomed2icd["idx_snomed"], df_snomed2icd["idx_icd"]))




all_mappings = {
    "icd2snomed": icd2snomed,
    "snomed2icd": snomed2icd,
    "id2snomed": id2snomed,
    "snomed2id": snomed2id,
    "id2icd": id2icd,
    "icd2id": icd2id,
}

with open(SAVE_PATH, "wb") as f:
    pickle.dump(all_mappings, f)

print(f"All mappings saved to: {SAVE_PATH}")

# embed all labels

from sentence_transformers import SentenceTransformer
device = torch.device("cuda" if torch.cuda.is_available() else "cpu")


snomed_info = pl.read_csv("D:/UMLS/snomed_info.csv")
icd_info = pl.read_csv("D:/UMLS/icd_info.csv")
with open(SAVE_PATH, "rb") as f:
    loaded_mappings = pickle.load(f)

# Access individual mappings like:
id2snomed = loaded_mappings["id2snomed"]
id2icd = loaded_mappings["id2icd"]
snomed2id = loaded_mappings["snomed2id"]
icd2id = loaded_mappings["icd2id"]
icd2snomed = loaded_mappings["icd2snomed"]
snomed2icd = loaded_mappings["snomed2icd"]

snomed_labels = snomed_info["SNOMED_label"].to_list()
icd_labels = icd_info["ICD_label"].to_list()

model = SentenceTransformer("yyzheng00/sapbert_lora_triplet_rank16_merged", device=device)
embeddings_snomed = model.encode(snomed_labels, batch_size=32, show_progress_bar=True)
embeddings_icd = model.encode(icd_labels, batch_size=32, show_progress_bar=True)

save_path_embed =  PATH + "sapbert_lora_triplet_rank16_merged.pt" 
torch.save({
    "model_name": "sapbert_lora_triplet_rank16",
    "snomed_embeddings": embeddings_snomed,
    "icd_embeddings": embeddings_icd,
}, save_path_embed)

embeddings_icd.shape
